requirements.txt
flask==2.0.3
flask_login==0.5.0
flask_migrate==2.7.0
flask_wtf==0.14.3
flask_sqlalchemy==2.4.4
sqlalchemy==1.3.23
email_validator==1.1.2
python-decouple==3.4
gunicorn==20.0.4
jinja2==3.0.3
psycopg2-binary==2.9.1
requests==2.27.1
selenium==4.0.0
celery==5.1.2
redis==3.5.3
babel==2.9.1
WTForms==3.0.0
proxycrawl==3.2.1
aiohttp==3.8.1
aiohttp_retry==2.4.6
Werkzeug==2.0.0
Flask-Cors==3.0.10
discord-webhook==0.16.3